Chargement OK du gros fichier excel (chargement et affichage assez lent)

+++ OtherProjects/LipoManager/LipoManager/LipoManager/Assets/QR_Scanner.cs
@@ -317,7 +317,7 @@ public class QR_Scanner : MonoBehaviour
 		public Element_Type Type = new Element_Type();
 		public string Name;
 
-		public int Courant;
+		public float Courant;
 		public int Capacite;
 		public int Restant;
 		public int DeltaV;
